Move thread decorator related classes to main library package to reduce classes visibility to default
